subscription admin api v1

+++ project/src/services/subscription.js
@@ -9,7 +9,14 @@ class SubscriptionService extends BaseService {
         DELETED: "subscription.deleted",
     }
 
-    constructor({manager, subscriptionRepository, subscriptionItemRepository, eventBusService, cartService, lineItemService}) {
+    constructor({
+                    manager,
+                    subscriptionRepository,
+                    subscriptionItemRepository,
+                    eventBusService,
+                    cartService,
+                    lineItemService
+                }) {
         super()
         this.manager_ = manager
         this.subscriptionRepository_ = subscriptionRepository
@@ -24,7 +31,15 @@ class SubscriptionService extends BaseService {
         return subscriptionRepo.find()
     }
 
-    async retrieve(subscriptionId, configs) {
+    async listAndCount(
+        config = {skip: 0, take: 50, order: {created_at: "DESC"}}
+    ) {
+        const subscriptionRepo = this.manager_.getCustomRepository(this.subscriptionRepository_)
+        const [data, total] = await subscriptionRepo.findAndCount(config)
+        return [data, total]
+    }
+
+    async retrieve(subscriptionId) {
         const subscriptionRepo = this.manager_.getCustomRepository(this.subscriptionRepository_)
         return subscriptionRepo.findOne(subscriptionId)
     }
@@ -46,13 +61,13 @@ class SubscriptionService extends BaseService {
 
             subscription.items = await Promise.all(
                 items.map(async (o) => {
-                    const res = subscriptionItemRepo.create({ ...o, subscription_id: subscription.id})
+                    const res = subscriptionItemRepo.create({...o, subscription_id: subscription.id})
                     await subscriptionItemRepo.save(res)
                     return res
                 })
             )
 
-            const result = await this.retrieve(subscription.id, { relations: ["variant", "items", "line_items"] })
+            const result = await this.retrieve(subscription.id, {relations: ["variant", "items", "line_items"]})
 
             await this.eventBus_
                 .withTransaction(manager)
@@ -65,13 +80,37 @@ class SubscriptionService extends BaseService {
     }
 
     async update(subscriptionId, data) {
-        const subscriptionRepo = this.manager_.getCustomRepository(this.subscriptionRepository_)
-        return subscriptionRepo.update(subscriptionId, data);
+        return this.atomicPhase_(async (manager) => {
+            const subscription = await this.retrieve(subscriptionId)
+            const subscriptionRepo = this.manager_.getCustomRepository(this.subscriptionRepository_)
+            const result = await subscriptionRepo.save(subscriptionRepo.merge(subscription, data))
+            await this.eventBus_
+                .withTransaction(manager)
+                .emit(SubscriptionService.Events.UPDATED, {
+                    id: subscriptionId,
+                })
+            return result
+        })
     }
 
-    async remove(subscriptionId){
-        const subscriptionRepo = this.manager_.getCustomRepository(this.subscriptionRepository_)
-        return subscriptionRepo.delete(subscriptionId);
+    async delete(subscriptionId) {
+        return this.atomicPhase_(async (manager) => {
+            const subscriptionRepo = this.manager_.getCustomRepository(this.subscriptionRepository_)
+
+            const subscription = await subscriptionRepo.findOne(subscriptionId)
+
+            if (!subscription) {
+                return Promise.resolve()
+            }
+            await subscriptionRepo.delete(subscriptionId);
+            await this.eventBus_
+                .withTransaction(manager)
+                .emit(SubscriptionService.Events.DELETED, {
+                    id: subscriptionId,
+                })
+
+            return Promise.resolve()
+        })
     }
 }
